ambientdisplay: Expose isDozing() to be used elsewhere
Previously the dream service would only report whether or not it was dreaming.
Ambient display counts as dreaming.  When dreaming, the volume inputs are
passed back to the application which is not what we want for ambient
display.

Fix this by exposing isDozing() and allow volume keys to control music
when we are dreaming and have moved into doze mode (aka, sleeping).

Change-Id: I3b1e72b95ec8b1b63ef0219259dfee77a5d339d3
Signed-off-by: Jason Edson <jaysonedson@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>