[android11-d1-release] More defensive checks for ProximitySensor
The bug this is fixing may already be resolved, but I wanted to
add a few more defensive checks to be sure. Also, I believe this
fixes a case where multiple listeners could receive different
proximity events if the first one pauses the proximity sensor before
the second listener is triggered.

Fixes: 156356906
Test: atest SystemUITests
Change-Id: Icdb23b8822f536d3c845c119509d180306fd8ad6
Signed-off-by: spezi77 <spezi7713@gmx.net>