[android11-d1-release] Keep USB disconnected receiver active when ADB activity in bg
When a new adb connection is attempted adbd will send the system's key
to the framework to prompt the user for consent to authorize the
connection. If another app is displayed over the adb prompt and the
user is not able to switch foreground apps (for instance during setup
wizard) the user will not be able to authorize the connection. This
commit keeps the USB disconnected receiver active even when the ADB
activity is in the background; if the user reseats the USB cable the
existing activity will be able to notify adbd and terminate, and a
new activity can be displayed in the foreground allowing the user to
authorize the connection.

Fixes: 152630548
Test: Manually verified the activity terminated and a new prompt was
      displayed after reseating the cable with the ADB activity in bg
Change-Id: I3293536639a20557982113ff9f6ac2bf6f7c1975
Signed-off-by: spezi77 <spezi7713@gmx.net>