cleanup: Tint active volume bar with gradient
On the commit:
https://github.com/Bootleggers-BrokenLab/frameworks_base/commit/8951e181ba9bccbed4b2b56fc1727622a740977c

A system to check and tint when the bool config_enableVolumePanelTint is enabled was implemented.
There was a typo of defining a bool that's already defined on the beginning. No need to double define it.
Also fix a logic that makes the volume panel background seekbar go gray when it should be a subtle accent tint

Signed-off-by: spezi77 <spezi7713@gmx.net>